Split up code for readability
